+++ source/isaaclab_assets/isaaclab_assets/evobot/balance/balance_env_cfg.py
@@ -83,9 +83,9 @@ class EvobotSceneConfig(InteractiveSceneCfg):
         gravity_bias=(0.0, 0.0, 0.0),
     )
 
-    # Contact sensor - mounted on head_link to detect illegal contacts
+    # Contact sensor - mounted on top_link to detect illegal contacts
     contact_forces_arm_link = ContactSensorCfg(
-        prim_path="/World/envs/env_.*/Robot/evobot/evobot/head_link",
+        prim_path="/World/envs/env_.*/Robot/evobot/evobot/top_link",
         update_period=0.01,
     )
 
@@ -98,8 +98,8 @@ class ActionCfg:
     - left_wheel_joint (Revolute)
     - right_wheel_joint (Revolute)
     - arm_joint (Revolute)
-    - left_grabbing_joint (Prismatic)
-    - right_grabbing_joint (Prismatic)
+    - left_gripper_joint (Prismatic)
+    - right_gripper_joint (Prismatic)
 
     All joints are at root level: /Robot/evobot/evobot/<joint_name>
 
@@ -132,8 +132,8 @@ class ActionCfg:
     grabber_effort = actions.JointEffortActionCfg(
         asset_name="robot",
         joint_names=[
-            "left_grabbing_joint",  # Prismatic - Left gripper
-            "right_grabbing_joint",  # Prismatic - Right gripper
+            "left_gripper_joint",  # Prismatic - Left gripper
+            "right_gripper_joint",  # Prismatic - Right gripper
         ],
         scale=100.0,  # Low force to avoid damaging objects
     )